Small GLFW changes
We needed several small GLFW changes in order to get things working.
We've added an error callback, which allowed us to see which other
things we needed; the GLFW_OPENGL_FORWARD_COMPAT was the result of
that.  And we always want GLFW_DOUBLEBUFFER.  Now the window appears
on OSX.  Hooray!