Scaling object parameters properly
We neglected to scale the position correctly upon receipt (or
assignment, at least), so that's now fixed.  Also, our randomly-
generated color was outside of [0.0, 1.0], so we rescaled things
to fall in that range.

Also a small tweak to configdata, since we no longer save the
password, and character name is more important, we fixed that.