Fixing leaks exposed by valgrind
We ran valgrind against the server program, and it uncovered some
important missteps.

The biggest hit was around 70k that we lost due to not calling
mysql_library_end().  Easy fix, big win.

We were also using an uninitialized variable to do an ID
comparison in the game object.

There was some questionable stuff in the action_rec, which keeps
track of the actions that the server knows about - the name
member was a char *, and we weren't handling things as cleanly
as possible.  Instead of worrying whether we've done the proper
memory management, we turned it into a std::string, and will let
the C++ memory manager handle it.

Last but not least, our hackery around the fake addrinfo structs
to let them also represent unix-domain sockets ended in hilarity.
We've now got a wrapper around freeaddrinfo to sense what we've
done, and handle it as expected.