Updates to r9client from the GL test program
Now that we were able to get pixels on the screen in the test
program, and have discovered some limitations of what we were
trying to do (and why it wasn't working), we've been able to get
the objects that are sent by the server to be displayed.

Our matrix math was incorrect, based on how the position and model
matrix interacted.  Now we're setting our rotations at the origin,
and *then* translating them into place.  Only then do we incorporate
the view and projection matrices.  We're still running into trouble
when we translate the camera away from the origin, but at least we
can see some stuff now.

One other item that we learned was that having a single VAO for
multiple buffers doesn't seem to work at all.  We've added a vao
element to the object structure, and are generating new VAOs for
each buffer.