Glyph cleanup
We had a bunch of unused and unneeded elements in our glyph
structure, so we cleaned them out.  The unit test now loads a
character, and does some testing of the various values.  Seems like
the font is close to done; we'll just need to get the label itself
to start using the font, and see how that works.