Add UTF-8 <--> uint32 string conversion
We will be sending everything over the wire in UTF-8, and we'll
need a way to get it into unsigned 32-bit Unicode integers for
display (Freetype uses unsigned longs, which seem to reference the
relevant Unicode code point).  So now we have functions to convert
both directions.

It's not clear whether this ought to be over in the protocol
library, or if it's fine here.