Start on rendering the string
We decided that since there is data available to the font that
doesn't need to be exposed to anybody else, that we'll generate
the actual string bitmap within the font object.  Finding the right
size is now private, and we've got a stub for rendering that does
very little.

Also updated the label itself to call the new font method.  It's
super clean.