Refine both the sizing and the rendering
We found some mathematical errors in the sizing method, and some
inconsistencies between the sizer and the renderer.  They look to
be arriving at the same size now, and putting things in the right
places within the buffer.  One of the main misses was that we were
totally disregarding the left element of the glyph within the
renderer, which made things look pretty bad in some places.