Fix incorrect uniform handling
We were using the attribute functions for getting the location of
our use_text uniform, which of course wasn't working.  Now that
we're getting the actual uniform location, we can set it correctly.

However, we found that when we were setting the uniform at buffer-
upload time, whichever the last label was to be set, set the
uniform for every label.  We've moved the uniform setting into the
draw method, and are also now unsetting it once we're done.  All
labels are now drawing as expected.