Add callback stubs for ui context
It seems like the most straightforward way to get our context to
hand out events is to set callbacks in it.  We're adding a set of
cursor callbacks here, one for motion, and one for button clicks.
We'll have to have care with the motion one, since we're going to
be doing position tracking in order to handle enter/leave events
for each of our children.

With general callbacks like the ones we're doing here, any
windowing system should be able to take advantage of our callbacks.