Add quadtree interface
We need to be able to quickly search for UI elements, and there is
nothing better for 2D sorting and searching than a quadtree.  This
interface is almost assuredly missing some important stuff, and
we've intentionally left some of the details out (the point class,
for example).  We might be able to use something from GLM for the
points.  The void pointers refer to a UI object, so we might just
use pointers to ui::panel objects.