Repair test Makefile for correct linkage
Travis exposed a couple of linking errors in the test directory's
Makefile.in.  A couple modules didn't have the correct set of
dependencies, and we weren't using SERVER_LDLIBS for the test
modules which needed it.

Also sneaking in a small fix to the POTFILES file - we don't need
to harvest anything from test programs.