Repair linking problems with client tests
When running the test programs on Linux, we found that linking with
the GLFW library, even though it wasn't being used, was causing
test failure.  We'll keep the CLIENT_LDLIBS variable within the
configure script, but we'll also produce XERCES_LDLIBS and
FREETYPE_LDLIBS variables, to allow proper linking within the unit
tests.