Repair font test by providing a .ttf file
We were counting on the fact that we could find a system font
called "Times New Roman.ttf", which wasn't the case on most Linux
hosts (it was a system font on OSX).  So we found a free font
called Techno Overload, and we're supplying it with the test suite.

http://all-free-download.com/font/download/techno_overload_font_4631.html