Push elements onto the front of the list
We'll carry the assumption that earlier in the list is further up
in the z stacking order.  To make that happen, we'll push things
onto the front of our list, rather than the back.

To raise something, we'll probably have to do a delete-add,
which is basically the same as when we move something.