Regenerate the search tree on context resize
Things will almost assuredly move between quadrants in our quadtree
when resizing the context, so we'll just take the simple step of
nuking the existing tree and completely regenerating it.  Resize
operations are not all that common, so having some complicated
(probably brittle) resizer, to save probably not much time, seems
misplaced.