Implement position and button callback drivers
We've got working searching, and working callback lists, so let's
put them all together.  We've got all 5 currently defined - enter,
leave, motion, button down and button up.

We save the current cursor position and the object (if any) the
cursor is pointing to.  This is so we can easily do the enter and
leave functions.