Convert geometry object to use GLM
For issue #81, we want to use GLM for both the server and client,
so we no longer need two dependencies.  Also, Eigen is showing a
whole bunch of warnings for register variables in Travis.  This
converts the geometry object to use GLM, which is the smallest
one we can start with.