Repair possible non-return state
In our socket creation function, we check for each of the possible
values that we expect, but if we get an unexpected value, we'll
fall out the bottom of the function without a return value.  We'll
do a dgram socket as our fallback, instead of explicitly checking
for that flag.