Repair handling of R-to-L and mixed-direction text
We made a lot of assumptions when we were first doing the label,
and not all of them were right.  We changed our is_l_to_r method
within the glyph to recognize various ranges of Unicode code points
which are typically R-to-L.  We also did some significant debugging
on the string renderer to be able to handle L-to-R, R-to-L, and
mixed-direction in both directions.

This will explicitly not handle the direction-setting Unicode code
points.  I'm not sure how we should go about those.  I think we'll
just ignore them for now, and if they come up, we'll worry about
them at that time.

Re: issue #77