Add call data structure for button events
While we're getting set up to make the text area, we need to be
able to tell where in the widget a button event occurs.  This will
allow us to place cursor events correctly, so we can identify
where the text cursor should go, or where the selection (if we
decide to do that) starts/ends.