Add cursor blink tracking in text field
We're adding a couple new members to the text field to track the
cursor's blinking.  We keep track of the time (of course), and
whether the cursor is visible at the moment.  Those C++ time
libraries are kind of crazy, and plenty (overly) verbose.

We're not drawing anything yet, but that's next on the slate.

Still cruising along on #72.