Add stubbed buffer population methods to text area
It's time to start trying to draw this monstrosity, so we'll add
some stubs for the eventual drawing.  We'll keep the main buffer
separate from the cursor, since we'll want to change the cursor's
visibility and/or position on a more regular basis, and we can do
only the regeneration that is necessary at the moment.