Start implementing the text area string generation
We'll do this in a couple stages.  First we'll get the basic panel
down, then add the properly-scaled string image, then we'll add the
cursor.  This is step one.

Setting the actual size of the widget didn't really make sense
in the max-size setter, so we're moving it to the buffer generation
method.