Resolve string resize errors in text field
We were seeing some string height changes in the text field,
depending on how high our ascenders or how low our descenders were.
First, the font bounding box descender was showing up as negative,
so we needed to negate it.  Second, we now take out the difference
between the full height and the actual string size, in order to
make scale properly.