Repair set-max-length call in UI test program
We do actually need a second arg when setting the maximum width of
the text area, so we'll add that to the UI test program.  Now it
actually changes the size as expected.

This should complete enough for #72 that we can merge this; we
still have a couple of TODO items, but this is good enough for now.