Delete all children when destroying the UI context
In the UI context, we want to consider all widgets that get passed
into us as ours, and we should blow them away once we're done with
them.  Since the widgets remove themselves from our children list
on destruction, all we need to do is delete them, and they'll take
care of the rest.