Use a temporary image for text field rendering
When the string is longer than can be displayed within the text
field borders, we want to keep the existing rendered image, but
we'll want to be able to rip out pieces of the image for display.
To that end, we'd like to use a temporary image while we render,
which should be deleted properly when it goes out of scope at the
end of the method.