Clean up string chunking in text field
The string chunking did a lot of calculations and loops that we
simply did not need to do.  We have all the data we need from the
couple calls, and we can do static calculations on those values.

If we scroll back through the string, the display doesn't change,
but that's more likely because none of the cursor motion actions
call into generate_string.