Add stub methods for child management in composite
We now have a child list and a set of methods to manage the child
list in the composite object.  The interface is copied almost
directly from the UI context.

I think what we'll do is change all the classes derived from panel
to either include a second constructor to use a composite as a
parent (can't call it parent, since that member already exists),
or even replace the constructors which take a context.  The
interface is the same, so it'll be just a question of renaming the
member we're using to access whichever object is the parent.

So we don't have to do a huge amount of refactor, I think we'll
also add a transformation uniform into our vertex shader, so we can
move the children around relative to the position of the composite.

Re: issue #78