Make the composite into a top-level class
The composite can (and should) be used for as much as it can be.
We noticed that in fleshing out the child management of the
composite, that we were copying the methods of the UI context
verbatim, so we decided that we would use the composite as more of
a "role" that can be inherited.  We'll factor the child management
out of the UI context next, and inherit the composite.