Add stub methods for child management in the manager
We want the ability of manager widgets to resize if they are either
too big or too small.  Overriding the child add/remove methods
seems to be the cleanest place to do that.

We'll eventually add some flags to control whether our widget
should be able to grow or shrink.

Re: issue #78