Add child spacing member to manager
We need the ability to space out the children in the manager
widget so that we can make things look nice.  Initially, it will be
used to space the children from the edges of the manager.
Eventually, in things like a row-column, it can be used to space
the child widgets apart.

Re: issue #78