Add method to ascertain correct size to manager
In the manager widget, we would like to wrap our children as
closely as possible, so we'll recompute size when we add or remove
any children.  We will eventually add flags to limit particular
size changes, e.g. we can allow only growth, only shrinking, both,
or neither.

Re: issue #78