Remove confusion in get/set size in manager
Multiple inheritance strikes!  We tried to override get/set size in
the row-column, and the compiler freaked out a little bit.  So
we're making our behaviour explicit here:  we favor the composite,
but we'll set both the composite and panel stuff (since we really
must).  Now we can override sizing in our descendents, and all will
be well.