Repair grid spilling in row-column
When we have a full grid size setting and too many elements for the
grid, we were getting incorrect values.  There were a couple errors
that we found.  First, we were setting the opposite dimensions in
our size setter - x corresponds with *columns*, and y corresponds
with *rows*, where we had it backwards.  Second, we were doing our
row-major spill in the x dimension, where that should have been the
column-major spill.