Clean up child-moving behaviour in manager
If we do a remove/add instead of just a move, this may change the
order in which our children are drawn or added to the widget, which
is absolutely not what we want to do.  Plus, moving just in the
tree and not also in the child lists will save us some unnecessary
work.