Clean up callback handling in row-column
We were not getting our size set properly in our parent.  It
turns out that we have two(!) size members, and we were neglecting
to set one of them... which was the one that the composite was
using to insert into its quadtree.

Also a change to our child-moving behaviour - the composite already
does the right thing, so we'll just do that.