Allow shared fonts in UI widgets
We were running into problems with the label and button widgets,
which all considered that they owned the font which they used,
when trying to get them to use a single shared font instance.  They
would all delete the font on destruction, which would cause
segfaults.  We now use the type argument into the font setter as a
flag which tells us if we're using a shared font, or the widget can
own it.

This is the entirety of issue #100.