Protect buttons from unchecked changes
When we attempted to set the buttons' active or armed states, we
didn't check whether the button was already active or armed, and
did the border-grow action regardless, possibly resulting in
incorrect border changes.  Now we'll only make a change if there
*was* a change.

While we were at it, when we deactivate the button, we'll also
disarm it.  Since we're no longer worried about unchecked changes,
this is safe.