Refactor UI elements to use GLM ivec2
We want to be able to set both width and height, or x- and y-
position in one shot per issue #98, so we want to at least take
ivec2 arguments to the get/set methods, but it makes more sense
to change the internal representation to ivec2 as well.

This is a double-sized commit, which is unfortunately necessary.
In order to refactor the panel, everything else under the panel
needs changes as well, so it can't be broken up meaningfully.