Implement callback routines in popup menu
We need to be able to show and hide the popup, so we've added some
callback routines.  In the constructor, we hook up to our parent,
whatever type it may be - we could have either a context or a
regular manager widget as a parent.  In our destructor, we unhook
from our parent.

Re: issue #66