Update UI text field size elements
In issue #114, we want to get rid of the special-case element of
'max_size', in favor of the regular 'size' in the text field
widget.  We add the max_width element in the ui::size namespace.
We can't just use ui::size::width, because manager widgets query
their children via the usual get interface to retrieve their
sizes, and the super-small numbers that the text field would
otherwise return are not accurate for the manager.

Additionally, we noticed some logic errors in the various get
methods, which were returning 1 on success, instead of the correct
0.