Add UI context to the client program
We add the UI context into the main function of the client program.
Since we can't assume that the client's shader program will always
be active, we'll add some calls to select it as we draw the objects
we've received from the server.

Also we'll make the resize function handle the UI context.

Re: issue #124