Repair incorrect child list handling in manager destructor
When we try to iterate over a list, and the list contents are
changed during our iteration, iterators become invalid.  We won't
use an iterator at all, but just see if there are objects, and
close the front one (which will remove the front one).

This does not complete issue #125, but was definitely a part of it.