Make sure saved child is valid in composite callback handlers
While tracking down the root cause of issue #125, we noticed that
the callback handlers in the composite could end up setting an
invalid object pointer as the object's saved pointer.  As we finish
each of the callbacks, we'll do another search to make sure we
still have a valid object.