Repair text field to correctly render its string
We were seeing some scaling behaviour in the text field as we
moved our cursor, and it turned out that we were always using the
left-of-cursor size when determining our actual string size.  Not
sure why this ever would have worked correctly, but it's working
once again.

This should finish up issue #124, so we can merge it.