Use monotonic clock for client cache purging
Along with issue #127, we found that gettimeofday() doesn't produce
monotonic results (leap seconds and DST changes will mess with the
clock results).  Also, we wanted something more portable, so we
turned to the C++11 std::chrono elements.  We tamed the verbosity
a bit with some typedefs.  Everything seems to be working smoothly.