Ensure that the subserver's client buffer is large enough
In issue #136, we noted a 'fixme' condition via Code Climate
analysis, and wrote a short test program to print out the sizes of
the various pieces of the protocol.  We noticed that we were
cramming entire in_addr/in6_addr structures into the server notice
packets, rather than keeping it simple and just using a textual IP
address.

The current size of the main packet union is 202 bytes, so the
1024-byte buffer will definitely fit anything that the protocol
would currently require.