Ensure dynamic casts fail appropriately
In issue #92, we had identified some possible instances in which
dynamic_cast calls might not have had their returns checked
appropriately.  We did find one instance in the main server which
might have failed, during system log cleanup.

There were some instances in the sockaddr family of classes which
can possibly throw exceptions due to casting a reference, but these
are all in constructors.  The factory function, contained in the
same file, can also throw exceptions for other reasons, so it's
reasonable to allow these instances to stand.  We did find some
blind casts in that same file, which we reworked to be reinterpret
casts, to match with the other objects.