Add member access method implementations in UI widget
The next set of implementation details to add is the access methods
for the new data members in the widget class.

We're going to try doing each of the boolean flags, which have no
subtype as ui::element::state, and have each flag as a subtype in
the ui::state namespace.

Re: issue #123