Update some virtualization in UI widget
We will almost certainly need to override the state access methods
in subclasses, so we'll go ahead and make those virtual.

We'll also use the override keyword to denote things we're actually
trying to override.  No functional changes, but adds a small bit of
safety, in case we get too crazy, or our parent decides to change
function signatures.

Re: issue #123