Resolve some errors with UI text field
We had apparently copy-pasted some code to deal with the cursor
VAO/VBO/EBO, and neglected to change the names of the EBO, so we
were getting segfaults when we tried to draw the cursor.

We also ensure that the element count for the EBO is initialized,
in case we try to draw the widget before we actually have anything
in the cursor buffers.

Lastly, we explicitly refer to the current object in the draw
method, when accessing the cursor transformation matrix.