Clean up UI manager methods
We were being a little sloppy with the set() method in the manager,
and had some stuff that didn't really belong there.  We'll move the
child handling for child spacing into the child spacing method.

Also, we ensure that we fall through to the appropriate superclass
for any of the resources that we don't explicitly handle.

Re: issue #123