Eliminate more lazy calling in UI row column
There was another case of our calling the set_size() method with
bogus parameters, instead of calling the right method.  A major
side-effect of this particular instance was that not all of the
children were being drawn correctly.  Now that the call is fixed,
the children are drawn correctly, and we are officially done with
this humongous refactor.