Repair UI text field string display
We were running into problems which we'd dealt with before, namely
that when the displayed string gained or lost a descender, it
was not taken into account when calculating the texture s/t values.
We can simply override and extend the label's generate_points()
method to slightly tweak its output.