Repair client logbuf errors found during testing
We found a couple of mistakes in the client logbuf.  We were not
using the right flags for opening - we should do std::ios::app
instead of std::ios::out (which truncates previous contents).
Also, we were closing the file on every trip through our loop,
which is very wrong.  Moving it ouside the loop makes everything
happy.