Add a resize callback list to the UI active
We need to be able to do things when resizes happen, mostly in the
context.  We'll add a new callback list for children to add things
into when there is a resize.  We'll also add a call data struct so
that the callees can get the new data without having to query it.

We've got no handler or function which calls this list yet; that
will be coming in a future commit.

Re: issue #64