Add a reference to the UI element in the client logbuf
We need to be able to associate an entry in the logbuf's list with
a UI widget, so that we can age things out as necessary; we have
the age in the logbuf, but not in the UI.  We don't do any of the
aging yet, mainly because I'm not sure of the best place to do it.

Also a small comparison cleanup:  we only care if our log window
pointer points to our log window, and if it's NULL, it doesn't
point to our window.

Re: issue #64