Only resize UI label when there are contents
We ran into a condition in the UI label in which it was trying to
split a string to a 0-width, and never returning.  It was due to
our having resized the widget before there were any contents to
make the widget nonzero-sized.  When doing the size recalculation,
we'll first make sure we've got some contents, and if not, we won't
do anything.