Repair possible divide-by-zero error in UI row-column
We had run into a condition when the column count was 0 and the
widget was set to column-major packing, that we would try to
divide by zero, and of course would generate an error.  We'll only
use the packing order when we're inserting items, and the
configured column count when we're deciding how big our grid should
be.