Repair cleanup thread in log display
The log display's cleanup thread had some significant problems,
which we have resolved.  We wait for the appropriate times, and
do cleanups in the appropriate way.  We still need to have the
widget set handle things in an asynchronous way sometimes, so that
will be handled in a forthcoming set of commits.