Add a namespace for sync/async child handling
We've been running into problems with close, depending on whether
we want to do things synchronously or asynchronously.  We're about
to add a flag (which will default to synchronous) that will change
close behaviour, in order to avoid more seg-faults.  Similarly,
adding/removing/moving children within composite widgets needs the
ability to defer changes until we're known to be in the correct
thread.  This commit adds a namespace and flag values for our
upcoming arguments.