Repair incorrect invocation of subserver program
We had been seeing a compiler warning for quite some time now
about an incorrect call to execlp().  Taking a look at the manpage,
we didn't need execlp() at all, since we were calling by full path.
execl() would do what we needed to do.

We also were passing incorrect parameters to the exec function;
the second arg needed to be the name of the binary we were running.