Remove sync flag from composite add_child method
It seems like our idea of having a flag whether to do things
synchronously or asynchronously was the wrong way to go.  We're
introducing loops that happen sometimes, and sometimes not, and
it has massively complicated everything.  So we're going to get
rid of the flag entirely, and handle all child changes
asynchronously.