Rework call path through composite child handling
We have removed all notion of synchronous child handling in the
composite widgets, so we need to refine the way our calls get
routed, to ensure that we make the right calls at the right times.

The draw method will now call out into the set_desired_size method,
instead of only clearing children queued for deletion.  We remove
the calls to set_desired_size from the queue-clearing method, to
remove looping.