Remove setting of dirty flag before child reflowing
When we perform the reinsertion of children into the row-column
widget, we had set the dirty flag as a signal to the child-handling
methods to not make any calls back into set_desired_size.  Now that
we will never make any such calls, we don't need to use the dirty
flag for this second(-ish) purpose any longer.