Reorder log entry modifications to ensure proper setting
We were setting full borders for our log entries, but only the
left/right were showing up.  We have some (maybe questionable) code
within the UI widget class which seeks to prevent setting of
borders when there is not sufficient size to support that border.
Because our multi-label widgets have a zero vertical size by
default, the top/bottom borders were not being set.  We reorder
the set calls to ensure that the widget has a size (due to having
a string as contents) before setting the border.

It may be the case that we should eventually override the border
and margin methods in the label widget, to turn this behaviour
off.