Allow the log cleanup thread to operate with disparate message sets
We found that once we'd cleaned up the last message from being
displayed in the log display widget, we'd never clean up another
one.

We need a lookahead function, but there's no general-purpose "add
X to an iterator" function.  We add one specific to our needs here,
and do a little refactor of the cleanup logic.