Add test for socket creation failure
We want to test the exceptional conditions within the client Comm
object, and unfortunately we'll need to mock out some of the
system functions that we rely on in the existing t_comm test.
So we'll make a separate file to test those exceptional conditions.
First up is a failure in the create_socket() method.