Add tests for exceptional conditions within client Comm start
The start() method attempts to start up a couple of threads, which
will generate exceptions upon failure.  We need to mock out all of
the pthreads library that we use, in order to detach completely
from actually doing any of the thread business.  We'll make sure
things are rolled back appropriately on failure.