Add exception checking for client configdata object
There are some situations in which we might throw exceptions, when
creating the directory structure for the client program.  We hadn't
tested them before, so here are some tests now.

In the course of adding these tests, we found that most of the
other tests depended on most of that structure being in place, so
we had to rework them as well.