Add test for unopened socket in server basesock
If a socket is not open, the object will have a 0 for the sock.
If we try to start up in this state, we should throw an exception.
It's going to be very difficult, if not impossible to get into this
state under normal operation, so this test might be mostly
pointless, but we're at least considering the case.