Ensure the max_id value is correct
While working on the unit test for the server GameObject, we
noticed that the max ID value may be wrong when setting to a
specific value.  So when we do get an actual value, we'll make
sure that our saved value is at least one higher than the value
we received.