Test exception handling within server thread pool startup
In the thread pool start() method, we can throw exceptions due to
pthread_create() failure, so that condition needs to be tested.

Also we noticed that if this occurs, we never unlock the pool
mutex, so we get put into a mostly-unrecoverable state.  We'll
unlock the mutex in our exception clause, to make sure we're at
least in a consistent state.

Re:  issue #152