Repair push/pop test for server thread pool
A std::string wasn't a good type to test the thread pool, due to
the fact that it maintains memory outside of the main object.
We'll add a simple type which keeps all its data in one chunk,
so we don't have to worry about pointers in our popped objects.