Repair lock-count discrepancies in server thread pool test
For some reason on our CI platform, we're seeing a whole bunch of
lock-unlock behaviour in the thread pool test, when there should
be none, so it seems like the test harness itself may be doing
some thread activity.  So instead of forcing lock counts to be
any specific number, we'll make sure they're nonzero, and the
lock and unlock counts are equal.