Remove the start() method from the action pool
In the writing of the unit test for the listening socket, we
noticed that there were some objects which were placed in very
questionable locations, notably the actions list was not placed
within the actions pool, even though that's the only thing that
uses it.

The result of the existing design is that if we want to test
anything related to the zone, we have to instantiate basically the
entire zone, and can't reasonably mock most of it.  This is an
obviously wrong design, so we'll fix it before continuing.

The changes are significant, and this will take several commits
to get done.