Move the execute_action() method into the action pool
Now that the action pool owns the actions map, there's no reason
for the zone to have anything to do with dispatching actions.
We'll move the execute_action() method where it needs to be.

We'll also keep a reference in the action pool to the game objects
map, which (for now) is maintained within the zone.