Pass an already-created library into the action pool constructor
In order to ease mocking of the action pool, we need to be able to
operate on a fake library, and the only way we can easily do that
is to accept a library object in the constructor.

Having to pass a library through the zone and into the action pool
is really rather ridiculous, and suggests that having the action
pool within the zone is the wrong place for it.  This suggests a
probable trend of decoupling things from the zone; the various
thread pools should move out of the zone entirely, and be objects
held within the scope of the main server program.