Add thread pools to the global scope in the server
We're moving the thread pools out of the zone, since they really
just don't belong there, and into the global scope of the server
program.

This will take several commits to complete, and hopefully we'll
be able to do things in a sensible way, without breaking the
working program along the way.  The tests will be broken as well,
since there will be significant changes required.