Add method to send ack for any given operation
In refactoring the action pool, we noticed that we were building
an ack packet inline, which is completely foolish.  The sending
socket should take care of constructing packets, and outside
callers should just make a simple call to cause it to happen.  So
we now have a send_ack method in the listening socket, which we
can also use in a couple instances in its descendents.