Simplify data flow in handling action requests
We've been doing most of the work of handling an action request
within the thread worker method, which is kind of a hassle to test.
We'll move that work into the execute_action method, so we can call
it directly in tests.

We now have the send_ack method within the listening sockets, so
we'll remove our construction of the ack packet in favor of calling
the helper method.