Reverse incorrect logic in action pool's execute_action() method
We were doing the test to see if the action request's source object
was actually valid for the control object that we have, and we had
the logic backward.  Previously we would only have done an action
if the object we were trying to make act *wasn't* owned by the
user in question.