Add test of unix socket unlink failure in server basesock
There was one clause we didn't have coverage on in the server
basesock object.  In the destructor, when the socket type is a
Unix domain socket, we try to unlink the socket file.  If the
unlink is unsuccessful, we log a message.  Unfortunately, our
CI platforms segfault unexplainably when we change the rdbuf
for std::clog; they don't on OSX, so we've added some #ifdefs
around that part.

Should have full coverage of the basesock now, even if we're
not always verifying everything.