Factor userid retrieval from listen socket login method
In writing the unit test for the server listen_socket's
login_user() method, we found a particular clause that could be
mostly factored out into its own method.  This has now become the
get_userid() method.

The test will be failing now, but will be resolved in the next
commit.