Refactor listen_socket and children
The login process was confusing, and had some duplicated code
that didn't ultimately need to be duplicated.  We pulled the
duplication into the listen_socket itself, and left the child
objects to create their specific types of users, which they now
pass *back* into the listen_socket for final handling.

This also has the side-effect of making the listen_socket a whole
lot easier to mock and test.