Simplify server database interface
During our refactor of the login process, we found that there was
a database method which didn't even need to exist, so we got rid
of it entirely.

We also found that in retrieving the character object ID for a
given character, we weren't taking the owner into account, so that
method now takes the userid as an additional argument.  The queries
in those methods are reworked to provide the right info.

We also fixed up the mock database object to reflect the new
changes in the interface.