Resolve build warning with STRING_UNKNOWN
In C++11, you can no longer legally use constant string literals
where a char * is expected.  The resolution is simple:  assign a
char array to the string in question, and use the array name in
place of the literal.  tcpd.h defines such a literal,
STRING_UNKNOWN, which we do use when making a hosts_ctl call, so
we add a char[] within the HAVE_LIBWRAP block at the top.

Also removed an unnecessary TODO comment; we have already done it.