Update UI font test to not mock out getenv
GTest apparently uses the environment to pass parameters around,
and trying to mock out getenv was just causing unrecoverable
failures.  Instead, we'll alter the actual environment of the
running test process, to coerce the behaviour we want to test.
Doesn't make me especially happy, but it does seem to work.