Convert redundant server console test to use factory functions
We don't need to test the same code twice with different args; we
test the basesock, which handles doing unix and inet sockets just
fine.  We'll instead test out the factory functions that we have
in our file.  They're simple and don't do much, but they're meant
to be very very simple.