Don't always delete console streambufs in server console session
We've been working on some testing for the server console, and we
have found that we may call the cleanup() method on the console
session more than once, so we should only do the cleanups once.
We'll use the socket as a flag to indicate whether the cleanup has
been done.