Ignore code that's not ours for purposes of Code Climate analysis
We don't care whether the libintl folks need to fix stuff, or if
somebody who wrote an M4 macro that's used within autotools wants
to revisit something.

At the moment, those are basically all of the items we actually
care about, so our analysis is going to be almost nothing until we
can get some of our other analyses going.